requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=20
pytest-django>=4.5
coverage